test(xds): move MES specs into clusters_test
